Tidy computeGroup into separate file
